app/requirements.txt
streamlit==1.29.0
openai==1.6.1
langchain==0.1.0
langchain-community==0.0.10
boto3==1.34.14
python-dotenv==1.0.0
